bin/requirements.txt
gps==3.22
pyserial==3.5
pyzmq==22.3.0
textual==0.32.0
influxdb-client==1.37.0
